+++ starships/src/main/scala/action/movement/VariableMovementActionBuilder.scala
@@ -1,5 +1,5 @@
 package action.movement
 
-case class VariableMovementActionBuilder(gradient: Int) extends MovementActionBuilder :
+case class VariableMovementActionBuilder(gradient: Double) extends MovementActionBuilder :
   def build(actionBuilderInput: MovementActionBuilderInput): VariableMovementAction =
     VariableMovementAction(gradient, actionBuilderInput.entity)
